HA: bedroom: don't handle "ghost" messages sent by the dimmer switch
The dimmer switch appears to send "ghost" messages (no action defined)
at regular intervals. I suppose this behavior is useful to record the
state of the dimmer switch, notably the remaining battery level.

Without this fix, the default behavior in
python_script.handle_hue_dimmer_switch is used, which is to turn on the
light. Let's say I'm tired of the light turning on by itself in the middle of
the night. 👀